requirements.txt
Flask~=3.1.0
Flask-SQLAlchemy
Flask-Migrate~=4.1.0
python-dotenv~=1.0.1
Flask-Cors~=5.0.1
pandas~=2.2.3
numpy~=2.2.3
pymysql
gunicorn
SQLAlchemy~=2.0.38
alembic~=1.14.1
requests~=2.32.3
scipy~=1.15.2